+++ R/forest.R
@@ -4361,6 +4361,8 @@ forest.meta <- function(x,
     o <- order(factor(x$bylevs, levels = bylevs))
     k.w <- x$k.w[o]
     k.w.hetstat <- k.w.hetstat[o]
+    if (x$hakn)
+      df.hakn.w <- x$df.hakn.w[o]
     ##
     TE.fixed.w <- x$TE.fixed.w[o]
     lower.fixed.w <- x$lower.fixed.w[o]
@@ -4415,6 +4417,8 @@ forest.meta <- function(x,
     else
       sel <- k.w > 0
     k.w <- k.w[sel]
+    df.hakn.w <- df.hakn.w[sel]
+    ##
     TE.fixed.w <- TE.fixed.w[sel]
     lower.fixed.w <- lower.fixed.w[sel]
     upper.fixed.w <- upper.fixed.w[sel]
@@ -5251,7 +5255,7 @@ forest.meta <- function(x,
                                 tt = statistics.effect.w[n.by + i],
                                 tp = rmSpace(pvals.effect.w[n.by + i],
                                              end = TRUE),
-                                df = k.w.hetstat - 1))
+                                df = df.hakn.w))
             else if (jama)
               text.effect.subgroup.random[[i]] <-
                 substitute(paste(tl,
@@ -5262,7 +5266,7 @@ forest.meta <- function(x,
                                 tt = statistics.effect.w[n.by + i],
                                 tp = rmSpace(pvals.effect.w[n.by + i],
                                              end = TRUE),
-                                df = k.w.hetstat - 1))
+                                df = df.hakn.w))
             else
               text.effect.subgroup.random[[i]] <-
                 substitute(paste(tl,
@@ -5273,7 +5277,7 @@ forest.meta <- function(x,
                                 tt = statistics.effect.w[n.by + i],
                                 tp = rmSpace(pvals.effect.w[n.by + i],
                                              end = TRUE),
-                                df = k.w.hetstat - 1))
+                                df = df.hakn.w))
           }
         }
         else {